Refactor bootloader OTA into ota_bootloader_esp_idf.cpp
Move the bootloader update flow out of the inline #ifdef blocks in
ota_backend_esp_idf.cpp and into a dedicated ota_bootloader_esp_idf.cpp,
mirroring the existing ota_partitions_esp_idf.cpp split. begin()/end()
shrink to one helper call per phase and the brace-outside-#ifdef pattern
in end() becomes a plain early-return for the bootloader path.

Also:
- Fix check_partition_access: error string was missing the f prefix so
  users saw the literal {option_string} placeholder.
- Reject an empty bootloader file with a clear error instead of letting
  data[0] raise IndexError.
- Document why handle_data_ now calls backend_->abort() unconditionally
  (begin() registers external partitions before esp_ota_begin).
- Add // USE_ESP32 / // USE_OTA_PARTITIONS end-comments in dump_config.
- Add unit tests for the empty-bootloader case and for --bootloader
  without allow_partition_access; strengthen the existing partition-table
  allow-flag test to assert the substituted flag name.
